Move settings-row to components and use button
