+++ src/panels/profile/ha-force-narrow-row.ts
@@ -9,7 +9,7 @@ import { fireEvent } from "../../common/dom/fire_event";
 import "../../components/ha-switch";
 import type { HaSwitch } from "../../components/ha-switch";
 import type { HomeAssistant } from "../../types";
-import "./ha-settings-row";
+import "../../components/ha-settings-row";
 
 @customElement("ha-force-narrow-row")
 class HaForcedNarrowRow extends LitElement {
