+++ src/panels/profile/ha-pick-theme-row.ts
@@ -12,7 +12,7 @@ import {
   css,
 } from "lit-element";
 import { HomeAssistant } from "../../types";
-import "./ha-settings-row";
+import "../../components/ha-settings-row";
 import { fireEvent } from "../../common/dom/fire_event";
 import "../../components/ha-formfield";
 import "../../components/ha-radio";
